+++ sdk/src/runtime/entries/auth.ts
@@ -1,1 +1,2 @@
+import "./types/worker";
 export * from "../lib/auth";
