+++ sdk/src/runtime/entries/client.ts
@@ -1,3 +1,4 @@
+import "./types/client";
 export * from "../client";
 export * from "../register/client";
 export * from "../lib/streams/consumeEventStream";
