+++ sdk/src/runtime/entries/clientSSR.ts
@@ -1,1 +1,2 @@
+import "./types/ssr";
 export * from "../lib/streams/consumeEventStream";
