+++ sdk/src/runtime/entries/router.ts
@@ -1,2 +1,3 @@
+import "./types/shared";
 export * from "../lib/router";
 export * from "../lib/links";
