+++ sdk/src/runtime/entries/ssr.ts
@@ -1,1 +1,2 @@
+import "./types/ssr";
 export * from "../register/ssr";
