+++ sdk/src/runtime/entries/types/client.ts
@@ -1,0 +1,1 @@
+import "./shared";
