+++ sdk/src/runtime/entries/types/shared.ts
@@ -1,0 +1,8 @@
+interface ImportMeta {
+  readonly env: ImportMetaEnv;
+}
+
+interface ImportMetaEnv {
+  readonly DEV: boolean;
+  readonly VITE_IS_DEV_SERVER: string;
+}
