+++ sdk/src/runtime/entries/types/ssr.ts
@@ -1,0 +1,1 @@
+import "./shared";
