+++ sdk/src/runtime/entries/types/worker.ts
@@ -1,0 +1,1 @@
+import "./shared";
