+++ sdk/src/runtime/worker.tsx
@@ -41,7 +41,10 @@ export const defineApp = <
         const url = new URL(request.url);
         url.pathname = url.pathname.slice("/assets/".length);
         return env.ASSETS.fetch(new Request(url.toString(), request));
-      } else if (IS_DEV && request.url.includes("/__vite_preamble__")) {
+      } else if (
+        import.meta.env.VITE_IS_DEV_SERVER &&
+        request.url.includes("/__vite_preamble__")
+      ) {
         return new Response(
           'import RefreshRuntime from "/@react-refresh"; RefreshRuntime.injectIntoGlobalHook(window); window.$RefreshReg$ = () => {}; window.$RefreshSig$ = () => (type) => type; window.__vite_plugin_react_preamble_installed__ = true;',
           {
