+++ sdk/tsconfig.json
@@ -43,6 +43,7 @@
 
     "types": [
       "./types/react.d.ts",
+      "./types/clientGlobal.d.ts",
       "./types/global.d.ts",
       "./types/vite.d.ts",
       "@cloudflare/workers-types"
